+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 # CHANGELOG
 
+## developing
+
+- New
+  - Importing Emoji Sequence data and new `EmojiSequence` class
+
+- Change:
+  - Re-structure the project
+  - Rename `EmojiData` class to `EmojiCharacter`
+  - Many other changes
+
+- Upgrade:
+  - Update emoji data files to 12.0
+
 ## 0.1.2
 
-- Date: 2019-01-010
+- Date: 2019-01-10
 
 - New
   - Sphinx documentations